requirements.txt
# ── Core Framework ─────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
websockets==13.1
python-multipart==0.0.20

# ── LLM — OpenAI ───────────────────────────────────────────
openai==1.58.1

# ── RAG / Document Processing ──────────────────────────────
langchain==0.3.13
langchain-openai==0.2.14
langchain-community==0.3.13
pymupdf==1.25.1           # PDF parsing (fitz)
python-docx==1.1.2        # DOCX parsing
beautifulsoup4==4.12.3    # HTML parsing
rank-bm25==0.2.2          # BM25 lexical search
sentence-transformers==3.3.1  # Cross-encoder re-ranking

# ── Database ───────────────────────────────────────────────
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
pgvector==0.3.6
alembic==1.14.0

# ── Cache / Queue ──────────────────────────────────────────
redis[asyncio]==5.2.1

# ── Voice — STT / TTS ──────────────────────────────────────
deepgram-sdk==3.9.0
pydub==0.25.1             # Audio conversion (mulaw ↔ PCM)
audioop-lts==0.2.1        # audioop para Python 3.13+ (stdlib removida en 3.13)
# elevenlabs==1.12.0     # Opcional: instalar solo si tts_provider=elevenlabs

# ── Telephony ──────────────────────────────────────────────
twilio==9.4.0

# ── Data / Config ──────────────────────────────────────────
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
httpx==0.28.1
aiofiles==24.1.0

# ── Security / Compliance ──────────────────────────────────
cryptography==44.0.0

# ── Language Detection ─────────────────────────────────────
langdetect==1.0.9
lingua-language-detector==2.1.0

# ── Observability ──────────────────────────────────────────
langfuse==2.57.4
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
prometheus-client==0.21.1

# ── Testing ────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1              # async test client

# ── Dev / Utilities ────────────────────────────────────────
loguru==0.7.3
tenacity==9.0.0            # retry logic
python-jose[cryptography]==3.3.0
